+++ packages/orq-rc/src/models/operations/createdatasource.ts
@@ -409,7 +409,7 @@ export const CreateDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KSYJC41M2D07PWE7VKCCDDXE"),
+  _id: z.string().default("01KT1Q4DVVGKQYVS8NYZMXWG43"),
   display_name: z.string(),
   description: z.string().optional(),
   status: CreateDatasourceStatus$inboundSchema,