+++ packages/orq-rc/src/models/operations/retrievedataset.ts
@@ -126,7 +126,7 @@ export const RetrieveDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-05-31T08:27:57.446Z",
+    "2026-06-01T13:48:56.269Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {