+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -3626,7 +3626,7 @@ export const Tools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   Tools
 > = z.object({
-  id: z.string().default("01KSYJC3KV18RWREHXRHHG8C9K"),
+  id: z.string().default("01KT1Q4DG57049RRGYBRQSKEKC"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),