+++ packages/orq-rc/RELEASES.md
@@ -9609,3 +9609,13 @@ Based on:
 - [typescript v4.10.0-rc.59] packages/orq-rc
 ### Releases
 - [NPM v4.10.0-rc.59] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.59 - packages/orq-rc
+
+## 2026-06-01 13:51:58
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.769.0 (2.892.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.10.0-rc.64] packages/orq-rc
+### Releases
+- [NPM v4.10.0-rc.64] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.64 - packages/orq-rc